[feat] 푸시메세지 FCM 스케쥴러 등록

+++ back/yourpilling/src/main/java/com/ssafy/yourpilling/push/model/service/vo/in/PushNotificationVo.java
@@ -9,7 +9,7 @@ import lombok.Value;
 @Builder
 public class PushNotificationVo {
 
-    Byte pushDay;
+    int pushDay;
     int hour;
     int minute;
